src/recognition/requirements.txt
pandas==2.2.3
Pillow==11.2.1
PyYAML==6.0.2
torch==2.2.0
torchvision==0.17.0